Change to single csp-demo program
