+++ Makefile
@@ -1,20 +1,15 @@
 LIBCSP       = ./libcsp
 LIBCSP_BUILD = ./libcsp/build
+LIBSOCKETCAN = ./libsocketcan
 BUILD        = ./build
 SHELL        = /bin/sh
 CC           = gcc
 RM           = rm -rf
 MKDIR        = mkdir -p
 
-csp-demo: clean server client
-
-server: server.c
+csp-demo: csp-demo.c
 	$(MKDIR) $(BUILD)
-	$(CC) -I $(LIBCSP)/include -I $(LIBCSP_BUILD)/include -pthread -o $(BUILD)/server server.c -L $(LIBCSP_BUILD) -l csp
-
-client: client.c
-	$(MKDIR) $(BUILD)
-	$(CC) -I $(LIBCSP)/include -I $(LIBCSP_BUILD)/include -pthread -o $(BUILD)/client client.c -L $(LIBCSP_BUILD) -l csp
+	$(CC) -I $(LIBCSP)/include -I $(LIBCSP_BUILD)/include -pthread -o $(BUILD)/csp-demo csp-demo.c -L $(LIBCSP_BUILD) -l csp
 
 .PHONY: clean
